Final project/requirements.txt
pytest
pytest-xdist
selenium
webdriver_manager
allure-pytest
sqlalchemy
bs4
lxml
pymysql
docker
